requirements.txt
langchain
langgraph
langchain-google-genai
langchain-community
google-search-results
Flask
Flask-Cors
python-dotenv